move Template::new() to Kernel::new()
