+++ src/Kernel.php
@@ -3,10 +3,30 @@ declare(strict_types=1);
 
 namespace Qiq;
 
+use Qiq\Compiler\Compiler;
+use Qiq\Compiler\QiqCompiler;
 use stdClass;
 
 abstract class Kernel
 {
+    public static function new(
+        string|array $paths = [],
+        string $extension = '.php',
+        string $encoding = 'utf-8',
+        string $cachePath = null,
+        HelperLocator $helperLocator = null,
+        Compiler $compiler = null,
+    ) : static
+    {
+        $helperLocator ??= HelperLocator::new(new Escape($encoding));
+        $compiler ??= new QiqCompiler($cachePath);
+
+        return new static(
+            new TemplateLocator((array) $paths, $extension, $compiler),
+            $helperLocator
+        );
+    }
+
     private Blocks $blocks;
 
     private string $content = '';
